+++ lib/rubocop/cop/rspec/mixin/css_selector.rb
@@ -5,59 +5,18 @@ module RuboCop
     module RSpec
       # Helps parsing css selector.
       module CssSelector
-        COMMON_OPTIONS = %w[
-          above below left_of right_of near count minimum maximum between text
-          id class style visible obscured exact exact_text normalize_ws match
-          wait filter_set focused
-        ].freeze
-        SPECIFIC_OPTIONS = {
-          'button' => (
-            COMMON_OPTIONS + %w[disabled name value title type]
-          ).freeze,
-          'link' => (
-            COMMON_OPTIONS + %w[href alt title download]
-          ).freeze,
-          'table' => (
-            COMMON_OPTIONS + %w[
-              caption with_cols cols with_rows rows
-            ]
-          ).freeze,
-          'select' => (
-            COMMON_OPTIONS + %w[
-              disabled name placeholder options enabled_options
-              disabled_options selected with_selected multiple with_options
-            ]
-          ).freeze,
-          'field' => (
-            COMMON_OPTIONS + %w[
-              checked unchecked disabled valid name placeholder
-              validation_message readonly with type multiple
-            ]
-          ).freeze
-        }.freeze
-        SPECIFIC_PSEUDO_CLASSES = %w[
-          not() disabled enabled checked unchecked
-        ].freeze
-
         module_function
 
-        # @param element [String]
-        # @param attribute [String]
-        # @return [Boolean]
+        # @param selector [String]
+        # @return [String]
         # @example
-        #   specific_pesudo_classes?('button', 'name') # => true
-        #   specific_pesudo_classes?('link', 'invalid') # => false
-        def specific_options?(element, attribute)
-          SPECIFIC_OPTIONS.fetch(element, []).include?(attribute)
-        end
+        #   id('#some-id') # => some-id
+        #   id('.some-class') # => nil
+        #   id('#some-id.cls') # => some-id
+        def id(selector)
+          return unless id?(selector)
 
-        # @param pseudo_class [String]
-        # @return [Boolean]
-        # @example
-        #   specific_pesudo_classes?('disabled') # => true
-        #   specific_pesudo_classes?('first-of-type') # => false
-        def specific_pesudo_classes?(pseudo_class)
-          SPECIFIC_PSEUDO_CLASSES.include?(pseudo_class)
+          selector.gsub(selector.scan(/[^\\]([>,+~.].*)/).join, '')
         end
 
         # @param selector [String]
@@ -67,6 +26,17 @@ module RuboCop
         #   id?('.some-class') # => false
         def id?(selector)
           selector.start_with?('#')
+        end
+
+        # @param selector [String]
+        # @return [Array<String>]
+        # @example
+        #   classes('#some-id') # => []
+        #   classes('.some-class') # => ['some-class']
+        #   classes('#some-id.some-class') # => ['some-class']
+        #   classes('#some-id.cls1.cls2') # => ['cls1', 'cls2']
+        def classes(selector)
+          selector.scan(/.([\w-]*)/).flatten
         end
 
         # @param selector [String]
@@ -87,19 +57,9 @@ module RuboCop
         def attributes(selector)
           selector.scan(/\[(.*?)\]/).flatten.to_h do |attr|
             key, value = attr.split('=')
+            value << ']' if value&.include?('[')
             [key, normalize_value(value)]
           end
-        end
-
-        # @param selector [String]
-        # @return [Boolean]
-        # @example
-        #   common_attributes?('a[focused]') # => true
-        #   common_attributes?('button[focused][visible]') # => true
-        #   common_attributes?('table[id=some-id]') # => true
-        #   common_attributes?('h1[invalid]') # => false
-        def common_attributes?(selector)
-          attributes(selector).keys.difference(COMMON_OPTIONS).none?
         end
 
         # @param selector [String]
@@ -122,7 +82,8 @@ module RuboCop
         #   multiple_selectors?('a.cls b#id') # => true
         #   multiple_selectors?('a.cls') # => false
         def multiple_selectors?(selector)
-          selector.match?(/[ >,+~]/)
+          normalize = selector.gsub(/(\\[>,+~]|\(.*\))/, '')
+          normalize.match?(/[ >,+~]/)
         end
 
         # @param value [String]
@@ -130,14 +91,14 @@ module RuboCop
         # @example
         #   normalize_value('true') # => true
         #   normalize_value('false') # => false
-        #   normalize_value(nil) # => false
+        #   normalize_value(nil) # => nil
         #   normalize_value("foo") # => "'foo'"
         def normalize_value(value)
           case value
           when 'true' then true
           when 'false' then false
-          when nil then true
-          else "'#{value}'"
+          when nil then nil
+          else "'#{value.gsub(/["']/, '')}'"
           end
         end
       end